erase as many features as possible
