--- src/app/(auth)/login/page.tsx
@@ -1,9 +1,0 @@
-import { SignIn } from "@clerk/nextjs";
-
-export default function Page() {
-  return (
-    <main className="flex min-h-screen items-center justify-center">
-      <SignIn />
-    </main>
-  );
-}
