--- src/app/(auth)/signup/page.tsx
@@ -1,9 +1,0 @@
-import { SignUp } from "@clerk/nextjs";
-
-export default function SignUpPage() {
-  return (
-    <main className="flex min-h-screen items-center justify-center">
-      <SignUp />
-    </main>
-  );
-}
