+++ src/app/api/route.ts
@@ -1,5 +1,5 @@
 import { NextResponse } from "next/server";
-import { api } from "@/experimental-json";
+import { api } from "@/v2";
 
 export async function GET() {
   return NextResponse.json(api);
