+++ src/app/globals.css
@@ -1,10 +1,3 @@
 @tailwind base;
 @tailwind components;
 @tailwind utilities;
-
-@layer base {
-  a,
-  button {
-    @apply select-none;
-  }
-}
