+++ src/app/page.tsx
@@ -1,144 +1,113 @@
-import { Open } from "@/icons";
-import Link from "next/link";
+"use client";
+import { Alert, Arrow, Check, Close, Zap } from "@/icons";
+import { useEffect, useState } from "react";
+import { propertyDictionary } from "@/v1";
+import { getRandomKey } from "@/utils";
+import { useEvaluation } from "@/hooks";
 
-export default function HomePage() {
+const initialState = {
+  correct: false,
+  incorrect: false,
+  partial: false,
+  notSubmitted: true,
+};
+
+export default function Home() {
+  const [attempt, setAttempt] = useState("");
+  const [cssProperty, setCssProperty] = useState("");
+  const {
+    state,
+    evaluateTranslation,
+    mutate: mutateTranslationStatus,
+  } = useEvaluation(initialState);
+  const { correct, incorrect, partial, notSubmitted } = state;
+
+  useEffect(() => {
+    setCssProperty(getRandomKey(propertyDictionary));
+  }, []);
+
+  const handleChange = (event: React.FormEvent<HTMLInputElement>) => {
+    const attempt = event.currentTarget.value;
+    setAttempt(attempt);
+  };
+
+  const resetInput = (afterMilisseconds = 800) => {
+    setTimeout(() => {
+      setCssProperty(getRandomKey(propertyDictionary));
+      setAttempt("");
+      mutateTranslationStatus({ type: "not_submitted" });
+    }, afterMilisseconds);
+  };
+
+  const handleKeyDown = (
+    event: React.KeyboardEvent<HTMLInputElement>,
+    translation: string
+  ) => {
+    if (event.code == "Enter") {
+      event.preventDefault();
+      const evaluation = evaluateTranslation(translation.trim(), cssProperty);
+
+      if (!evaluation) return;
+
+      resetInput();
+    }
+  };
+
   return (
-    <main className="flex h-full w-full flex-col justify-center divide-y-2 divide-zinc-700 text-center">
-      <section
-        id="hero-banner"
-        className="mx-auto flex max-w-full flex-col items-center justify-center gap-10 px-2 py-20 sm:max-w-7xl md:py-32"
-      >
-        <h1 className="break-words text-5xl font-bold text-white md:text-7xl">
-          Learn
-          <span className="text-berryBlue"> Tailwind </span>
-          by playing
-        </h1>
-        <p className="text-md px-6 text-zinc-400 md:text-xl">
-          Dive into the most ergonomic styling framework and{" "}
-          <strong className="text-white">boost your development </strong>
-          speed through{" "}
-          <strong className="text-white">gamified learning</strong>.
-        </p>
-        <div className="flex flex-col items-center gap-6 md:flex-row md:gap-12">
-          <Link
-            className="text-md w-52 border border-berryBlue bg-inherit py-2 font-medium text-berryBlue underline-offset-4 transition-all hover:underline md:w-auto md:px-6 md:py-4 md:text-xl"
-            href="/play"
-          >
-            Play Now
-          </Link>
-          <Link
-            className="text-md flex w-52 items-center justify-center gap-4 border border-zinc-700 py-2 transition-all hover:border-white md:w-auto md:px-6 md:py-4 md:text-xl"
-            href="/help"
-          >
-            Learn More
-            <Open className="fill-white" />
-          </Link>
-        </div>
-      </section>
-      <section
-        id="benefits"
-        className="mx-auto flex max-w-full flex-col items-center justify-center gap-16 border-zinc-700 px-2 py-20 sm:max-w-7xl md:py-32"
-      >
-        <div className="flex flex-col items-center gap-6">
-          <span className="text-lg italic tracking-wider">
-            <span className="text-white">Noob?</span>{" "}
-            <span className="text-berryBlue">User?</span>{" "}
-            <span className="text-purplePlus">Adept?</span>
-          </span>
-          <h2 className="break-words text-5xl font-bold text-white md:text-6xl">
-            Explore. <span className="text-berryBlue">Practice. </span>
-            <span className="text-purplePlus">Master. </span>
-          </h2>
-        </div>
-        <div className="flex w-full flex-col items-center gap-16 md:flex-row">
-          <div className="flex flex-col gap-6 text-center md:text-left">
-            <h3 className="text-2xl font-medium text-white md:text-4xl">
-              Active Learning
-            </h3>
-            <p className="text-md text-zinc-400 md:text-lg">
-              Games empower players with active decision-making,
-              problem-solving, and real-time knowledge application, fostering
-              deep understanding and concept retention.
-            </p>
+    <main className="flex flex-col justify-center divide-y-2 divide-zinc-700 text-center">
+      <section className="mx-auto flex max-w-7xl flex-col items-center gap-10 py-[268px]">
+        <div className="flex items-center justify-center gap-6">
+          <div className="flex cursor-default flex-col items-start gap-2 text-zinc-400">
+            .class &#123;
+            <span
+              className={`
+                ${
+                  notSubmitted &&
+                  "border border-berryBlue font-medium text-berryBlue"
+                } 
+                ${correct && "border border-greenGo font-medium text-greenGo"}
+                ${
+                  partial &&
+                  "border border-yellowYes font-medium text-yellowYes"
+                }
+                ${
+                  incorrect &&
+                  "animate-shake border border-alertRed font-medium text-alertRed"
+                }
+                w-96 origin-center select-all bg-transparent p-5 text-xl transition-all`}
+            >
+              {cssProperty}
+            </span>
+            &#125;
           </div>
-          <div className="flex flex-col gap-6 text-center md:text-left">
-            <h3 className="text-2xl font-medium text-white md:text-4xl">
-              Immediate Feedback
-            </h3>
-            <p className="text-md text-zinc-400 md:text-lg">
-              Active games foster decision-making, problem-solving, real-time
-              knowledge application, and deeper understanding with concept
-              retention.
-            </p>
+          {notSubmitted && <Arrow size={32} className="fill-white" />}
+          {correct && <Check size={32} className="fill-greenGo" />}
+          {partial && <Alert size={32} className="fill-yellowYes" />}
+          {incorrect && (
+            <Close
+              size={32}
+              className="animate-shake fill-alertRed transition-all"
+            />
+          )}
+          <div className="flex cursor-default flex-col items-start gap-2 text-zinc-400">
+            className=&#34;
+            <input
+              value={attempt}
+              onChange={(event) => handleChange(event)}
+              onKeyDown={(event) => handleKeyDown(event, attempt)}
+              autoFocus
+              className={`transition-al w-96 origin-center bg-transparent p-5 text-xl focus:outline-none
+              ${notSubmitted && "border border-white font-medium text-white"}
+              ${correct && "border border-greenGo font-medium text-greenGo"}
+              ${partial && "border border-yellowYes font-medium text-yellowYes"}
+              ${
+                incorrect &&
+                "animate-shake border border-alertRed font-medium text-alertRed"
+              }`}
+            />
+            &#34;
           </div>
         </div>
-      </section>
-      <section
-        id="faq"
-        className="mx-auto flex w-full max-w-full flex-col gap-16 px-2 py-20 sm:max-w-7xl md:py-32"
-      >
-        <div className="flex flex-col items-center gap-6">
-          <p className="text-lg italic tracking-wider">
-            Have doubts?
-            <span className="text-berryBlue"> Good!</span>
-          </p>
-          <h2 className="break-words text-5xl font-bold text-white md:text-6xl">
-            Frequently Asked <span className="text-berryBlue">Questions</span>
-          </h2>
-        </div>
-        <ul className="flex max-w-3xl flex-col gap-6 self-center">
-          <li className="flex flex-col gap-3 px-2 text-center md:px-0 md:text-left">
-            <p className="text-lg font-medium md:text-2xl">
-              What is <span className="text-berryBlue">Tailwind?</span>
-            </p>
-            <p className="text-sm text-zinc-400 md:text-lg">
-              TailwindCSS is a utility-first framework with pre-built CSS
-              classes for responsive, customizable UIs. It promotes a functional
-              styling approach, offering flexibility through direct class
-              manipulation for efficient design control.
-            </p>
-          </li>
-          <li className="flex flex-col gap-3 px-2 text-center md:px-0 md:text-left">
-            <p className="text-lg font-medium md:text-2xl">
-              Isn&apos;t that just
-              <span className="text-berryBlue"> inline styles?</span>
-            </p>
-            <p className="text-sm text-zinc-400 md:text-lg">
-              TailwindCSS goes beyond inline styles by providing a predefined
-              design system, allowing for consistent UIs with constraints. It
-              offers responsive utilities for building fully responsive
-              interfaces, unlike inline styles. Tailwind&apos;s state variants
-              enable styling hover, focus, and other states not possible with
-              inline styles.
-            </p>
-          </li>
-          <li className="flex flex-col gap-3 px-2 text-center md:px-0 md:text-left">
-            <p className="text-lg font-medium md:text-2xl">
-              Why not use <span className="text-berryBlue">Bootstrap </span>
-              then?
-            </p>
-            <p className="text-sm text-zinc-400 md:text-lg">
-              TailwindCSS uses utility classes directly in HTML for styling
-              components, providing flexibility and customization without
-              overriding styles or writing custom CSS. It diverges from
-              Bootstrap&apos;s pre-defined components, embracing a utility-first
-              approach for efficient styling.
-            </p>
-          </li>
-          <li className="flex flex-col gap-3 px-2 text-center md:px-0 md:text-left">
-            <p className="text-lg font-medium md:text-2xl">
-              Doesn&apos;t the
-              <span className="text-berryBlue"> HTML get ugly?</span>
-            </p>
-            <p className="text-sm text-zinc-400 md:text-lg">
-              TailwindCSS&apos;s utility classes may increase HTML verbosity,
-              but the benefits of scalability, maintainability, rapid
-              development, and flexibility outweigh concerns about
-              &quot;ugliness.&quot; It offers a functional approach to styling
-              and reduces the need for custom CSS.
-            </p>
-          </li>
-        </ul>
       </section>
     </main>
   );
