--- src/icons/Open.tsx
@@ -1,15 +1,0 @@
-import { IconProps } from "./IconProps.types";
-
-export function Open({ className, size = 24 }: IconProps) {
-  return (
-    <svg
-      xmlns="http://www.w3.org/2000/svg"
-      width={size}
-      height={size}
-      viewBox="0 0 24 24"
-      className={className}
-    >
-      <path d="M21 11V3h-8v2h4v2h-2v2h-2v2h-2v2H9v2h2v-2h2v-2h2V9h2V7h2v4h2zM11 5H3v16h16v-8h-2v6H5V7h6V5z" />
-    </svg>
-  );
-}
