+++ src/icons/index.ts
@@ -1,16 +1,8 @@
 export { Zap } from "./Zap";
 export { Check } from "./Check";
-export { DoubleCheck } from "./DoubleCheck";
 export { Close } from "./Close";
-export { Copy } from "./Copy";
 export { Arrow } from "./Arrow";
-export { Login } from "./Login";
-export { Open } from "./Open";
 export { House } from "./House";
 export { Info } from "./Info";
-export { Restart } from "./Restart";
 export { Alert } from "./Alert";
-export { Swap } from "./Swap";
-export { Trend } from "./Trend";
 export { Menu } from "./Menu";
-export { MultipleNotes } from "./MultipleNotes";
