--- src/middleware.ts
@@ -1,7 +1,0 @@
-import { authMiddleware } from "@clerk/nextjs";
-
-export default authMiddleware({ publicRoutes: ["/", "/api", "/help"] });
-
-export const config = {
-  matcher: ["/((?!.*\\..*|_next).*)", "/", "/(api|trpc)(.*)"],
-};
